developer and databaseadmin now pass tests

+++ src/main/java/staff/management/Manager.java
@@ -5,7 +5,7 @@ import staff.Employee;
 public class Manager extends Employee {
 
     private String deptName;
-    public Manager(String name, String nationalInsuranceNumber, int salary, String deptName)
+    public Manager(String name, String nationalInsuranceNumber, double salary, String deptName)
     {
         super(name, nationalInsuranceNumber, salary);
         this.deptName = deptName;
